Moved todos to /api/. Added example to Dashboard.vue
